+++ packages/abap-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @sap-ux/abap-deploy-config-sub-generator
 
+## 0.0.71
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.2.50
+-   @sap-ux/deploy-config-generator-shared@0.0.46
+
+## 0.0.70
+
+### Patch Changes
+
+-   39cf280: apply validation for package format and starting prefix with flag
+-   Updated dependencies [39cf280]
+    -   @sap-ux/abap-deploy-config-inquirer@1.2.49
+
+## 0.0.69
+
+### Patch Changes
+
+-   @sap-ux/abap-deploy-config-inquirer@1.2.48
+-   @sap-ux/deploy-config-generator-shared@0.0.46
+-   @sap-ux/abap-deploy-config-writer@0.0.107
+
 ## 0.0.68
 
 ### Patch Changes