+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/app-config-writer
 
+## 0.5.51
+
+### Patch Changes
+
+-   e32dd29: outsource methods for reuse
+
+## 0.5.50
+
+### Patch Changes
+
+-   Updated dependencies [a3a43b2]
+    -   @sap-ux/axios-extension@1.20.1
+
 ## 0.5.49
 
 ### Patch Changes