+++ packages/cf-deploy-config-writer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.2.12
+
+### Patch Changes
+
+-   b339248: Ensure missing package.json id matches the mta ID
+
 ## 0.2.11
 
 ### Patch Changes