+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/deploy-tooling
 
+## 0.15.58
+
+### Patch Changes
+
+-   Updated dependencies [a3a43b2]
+    -   @sap-ux/axios-extension@1.20.1
+    -   @sap-ux/system-access@0.5.37
+
 ## 0.15.57
 
 ### Patch Changes