+++ packages/fiori-app-sub-generator/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @sap-ux/fiori-app-sub-generator
 
+## 0.0.31
+
+### Patch Changes
+
+-   Updated dependencies [2050952]
+    -   @sap-ux/ui5-info@0.10.1
+    -   @sap-ux/launch-config@0.9.5
+    -   @sap-ux/ui5-application-inquirer@0.12.4
+    -   @sap-ux/odata-service-inquirer@2.3.9
+
+## 0.0.30
+
+### Patch Changes
+
+-   38519fa: Fix for cached connection crash
+-   Updated dependencies [38519fa]
+    -   @sap-ux/odata-service-inquirer@2.3.8
+
+## 0.0.29
+
+### Patch Changes
+
+-   0d8918a: updates for cap and virtual endpoints
+-   Updated dependencies [0d8918a]
+    -   @sap-ux/ui5-application-inquirer@0.12.3
+    -   @sap-ux/fiori-elements-writer@2.3.6
+    -   @sap-ux/fiori-freestyle-writer@2.3.4
+    -   @sap-ux/odata-service-inquirer@2.3.7
+
+## 0.0.28
+
+### Patch Changes
+
+-   Updated dependencies [8d4bca6]
+    -   @sap-ux/odata-service-writer@0.26.17
+    -   @sap-ux/fiori-elements-writer@2.3.5
+    -   @sap-ux/fiori-freestyle-writer@2.3.3
+    -   @sap-ux/odata-service-inquirer@2.3.7
+
+## 0.0.27
+
+### Patch Changes
+
+-   Updated dependencies [a3a43b2]
+    -   @sap-ux/axios-extension@1.20.1
+    -   @sap-ux/odata-service-inquirer@2.3.7
+
 ## 0.0.26
 
 ### Patch Changes