+++ packages/flp-config-inquirer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/flp-config-inquirer
 
+## 0.2.84
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.41
+-   @sap-ux/adp-tooling@0.13.39
+
+## 0.2.83
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.38
+
+## 0.2.82
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.37
+
 ## 0.2.81
 
 ### Patch Changes