+++ packages/ui-service-sub-generator/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @sap-ux/ui-service-sub-generator
 
+## 0.0.34
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.6.41
+-   @sap-ux/ui-service-inquirer@0.0.34
+
+## 0.0.33
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.33
+
+## 0.0.32
+
+### Patch Changes
+
+-   @sap-ux/ui-service-inquirer@0.0.32
+
+## 0.0.31
+
+### Patch Changes
+
+-   Updated dependencies [a3a43b2]
+    -   @sap-ux/axios-extension@1.20.1
+    -   @sap-ux/system-access@0.5.37
+    -   @sap-ux/ui-service-inquirer@0.0.31
+
 ## 0.0.30
 
 ### Patch Changes